+++ database/migrations/1673786935770_task_partners.ts
@@ -5,10 +5,10 @@ export default class extends BaseSchema {
 
     public async up() {
         this.schema.createTable(this.tableName, (table) => {
-            table.increments('id')
+            table.bigIncrements('id')
 
-            table.integer('task_id').unsigned().references('id').inTable('tasks').onDelete('CASCADE')
-            table.integer('user_id').unsigned().references('id').inTable('users').onDelete('CASCADE')
+            table.bigInteger('task_id').unsigned().references('id').inTable('tasks').onDelete('CASCADE')
+            table.bigInteger('user_id').unsigned().references('id').inTable('users').onDelete('CASCADE')
 
             // For MySQL
             // table.timestamps(true, true)
